requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
google-genai
pypdf2==3.0.1
pandas==2.1.4
openpyxl==3.1.5
pytest==7.4.3
